+++ TesteWeatherApi/TesteWeatherApi/Interfaces/Repositories/IBaseRepository.cs
@@ -3,7 +3,7 @@
     public interface IBaseRepository<Entity>
     {
         Task<Entity> Get(int id);
-        Task<Entity> Update(Entity entity, long id);
+        Task<Entity> Update(Entity entity, int id);
         Task<bool> Delete(int id);
         Task<Entity> Create(Entity entity);
         Task<List<Entity>> GetAll();
